requirements.txt
django
pillow
django-gravatar2
django-markdownx
django-markdown-deux
